concurrent option is defined similar to option
The ConcurrentOption is initially designed with an api as similar as possible to the standard Option.
* methods requiring &mut self and self are identical
* methods requiring &self require an additional parameter for the Ordering

Almost all common standard option methods are covered. In this way, it is not very useful, since it just replicates the standard option with additional overhead. Specialized concurrent methods will be implemented separately to define the main useful capabilities of the type.