insert_if_none is implemented
Insert if none is implemented as the first concurrent method. The method allows to insert a value into the option if it was none. The method can be called concurrently with a shared reference.

Correct orderings for the write operation is defined and tested. However, reader methods should also pick an ordering stronger than Relaxed if a concurrent write operation is happening. This is added as a documentation.

The method safely allows multiple readers and multiple writers, as defined in the tests.